adding some final logics and cleaning the code

+++ calories/models.py
@@ -20,7 +20,7 @@ class Profile(models.Model):
 	person_of = models.ForeignKey(User,null=True,on_delete=models.CASCADE)
 	calorie_count = models.FloatField(default=0,null=True,blank=True)
 	food_selected = models.ForeignKey(Food,on_delete=models.CASCADE,null=True,blank=True)
-	quantity = models.PositiveIntegerField(default=0)
+	quantity = models.FloatField(default=0)
 	total_calorie = models.FloatField(default=0,null=True)
 	date = models.DateField(auto_now_add = True)
 	calorie_goal = models.PositiveIntegerField(default=0)
@@ -52,5 +52,4 @@ class PostFood(models.Model):
     profile = models.ForeignKey(Profile,on_delete=models.CASCADE)
     food = models.ForeignKey(Food,on_delete=models.CASCADE)
     calorie_amount = models.FloatField(default=0,null=True,blank=True)
-    amount = models.PositiveIntegerField(default=0)
-
+    amount = models.FloatField(default=0)
